Updated underlying container of ValueContainer to be based on std::tuple rather than std::array. Removes need for std::variant, much more memory efficient
