+++ include/CLArgs/value_container.hpp
@@ -3,58 +3,14 @@
 
 #include <CLArgs/core.hpp>
 
-#include <cstdint>
+#include <cstddef>
 #include <optional>
 #include <tuple>
+#include <type_traits>
 #include <variant>
 
 namespace CLArgs
 {
-    template <typename T, typename Variant>
-    struct PrependTypeToVariant;
-
-    template <typename T, typename... Ts>
-    struct PrependTypeToVariant<T, std::variant<Ts...>>
-    {
-        using Type = std::variant<T, Ts...>;
-    };
-
-    template <typename... Ts>
-    struct UniquelyFilteredVariantImpl;
-
-    template <typename T, typename... Ts>
-    struct UniquelyFilteredVariantImpl<T, Ts...>
-    {
-        using Type = std::conditional_t<IsPartOf<T, Ts...>,
-                                        typename UniquelyFilteredVariantImpl<Ts...>::Type,
-                                        typename PrependTypeToVariant<T, typename UniquelyFilteredVariantImpl<Ts...>::Type>::Type>;
-    };
-
-    template <>
-    struct UniquelyFilteredVariantImpl<>
-    {
-        using Type = std::variant<>;
-    };
-
-    template <typename... Ts>
-    using UniquelyFilteredVariant = typename UniquelyFilteredVariantImpl<Ts...>::Type;
-
-    template <typename T, typename Tuple>
-    struct TupleTypeIndex;
-
-    template <typename T, typename... Types>
-    struct TupleTypeIndex<T, std::tuple<T, Types...>>
-    {
-        static constexpr std::size_t value{0};
-    };
-
-    template <typename T, typename U, typename... Types>
-    struct TupleTypeIndex<T, std::tuple<U, Types...>>
-    {
-        static constexpr std::size_t value = 1 + TupleTypeIndex<T, std::tuple<Types...>>::value;
-    };
-
-
     template <Parsable... Parsables>
     class ValueContainer
     {
@@ -71,14 +27,15 @@ namespace CLArgs
         template <Parsable T>
         static consteval std::size_t index_of_type();
 
-        std::array<UniquelyFilteredVariant<std::optional<typename Parsables::ValueType>...>, sizeof...(Parsables)> values_;
+        using ValuesTuple = std::tuple<std::optional<typename Parsables::ValueType>...>;
+        static_assert(AllUnique_v<Parsables...>, "Duplicate template parameter types is not allowed in ValueContainer");
+        ValuesTuple values_;
     };
 } // namespace CLArgs
 
 template <CLArgs::Parsable... Parsables>
-CLArgs::ValueContainer<Parsables...>::ValueContainer()
+CLArgs::ValueContainer<Parsables...>::ValueContainer() : values_{std::make_tuple(std::optional<typename Parsables::ValueType>{}...)}
 {
-    ((values_[index_of_type<Parsables>()] = std::optional<typename Parsables::ValueType>{std::nullopt}), ...);
 }
 
 template <CLArgs::Parsable... Parsables>
@@ -87,7 +44,8 @@ void
 CLArgs::ValueContainer<Parsables...>::set_value(const typename T::ValueType &value)
 {
     constexpr std::size_t index = index_of_type<T>();
-    values_[index]              = std::optional<typename T::ValueType>{value};
+    static_assert(std::is_same_v<std::tuple_element_t<index, ValuesTuple>, std::optional<typename T::ValueType>>);
+    std::get<index>(values_) = std::optional<typename T::ValueType>{value};
 }
 
 template <CLArgs::Parsable... Parsables>
@@ -96,7 +54,8 @@ const std::optional<typename T::ValueType> &
 CLArgs::ValueContainer<Parsables...>::get_value() const
 {
     constexpr std::size_t index = index_of_type<T>();
-    return std::get<std::optional<typename T::ValueType>>(values_[index]);
+    static_assert(std::is_same_v<std::tuple_element_t<index, ValuesTuple>, std::optional<typename T::ValueType>>);
+    return std::get<index>(values_);
 }
 
 template <CLArgs::Parsable... Parsables>
@@ -104,7 +63,7 @@ template <CLArgs::Parsable T>
 consteval std::size_t
 CLArgs::ValueContainer<Parsables...>::index_of_type()
 {
-    return TupleTypeIndex<T, std::tuple<Parsables...>>::value;
+    return TupleTypeIndex_v<T, std::tuple<Parsables...>>;
 }
 
 #endif
